+++ lib/momentum.dart
@@ -14,4 +14,6 @@ export 'src/momentum_base.dart'
         InjectService,
         MomentumTester,
         RouterMixin;
-export 'src/momentum_router.dart' show Router, RouterPage, RouterParam;
+export 'src/momentum_router.dart' show MomentumRouter, RouterPage, RouterParam;
+// ignore: deprecated_member_use_from_same_package
+export 'src/momentum_router.old.dart' show Router;